requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
numba>=0.56.0
ruamel.yaml>=0.17.0
requests>=2.25.0
pyserial>=3.5
crcmod>=1.7
Pillow>=8.0.0
scipy>=1.7.0